backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-magic==0.4.27
edge-tts==6.1.9
emails==0.6.0
jinja2==3.1.2 